remove files sources and separate some logic
