+++ alerts.example.config
@@ -2,9 +2,6 @@
 # GLOBAL SETTINGS
 #------------------------------------------------------------------------------
 
-# Can either be "files" or "database", nothing else.
-# alerts_source = "files"
-alerts_path = ".alerts/"
 wss_domain = "yourpgcdc.instance.cloud"
 
 #------------------------------------------------------------------------------
