+++ src/main.rs
@@ -1,37 +1,31 @@
 #[macro_use]
 extern crate log;
 
-use crate::sp_alerts::SpAlerts;
 use crate::utils::config::Config;
+use crate::utils::monitor::Monitor;
 
 use ahash::AHashMap;
-use clap::{Parser, Subcommand};
+use clap::Parser;
 use diesel::{prelude::PgConnection, r2d2::ConnectionManager};
-use futures::FutureExt;
 use sproot::models::AlertsConfig;
-use sproot::prog;
+use sproot::{prog, Pool};
+use websockets::ws_handler::WsHandler;
+use websockets::ws_message::{msg_err_handler, msg_ok_database};
 use std::sync::RwLock;
+use std::{thread, time::Duration};
+use utils::mail::test_smtp_transport;
 
-mod check;
-mod sp_alerts;
 mod utils;
+mod websockets;
 
 #[derive(Parser, Debug)]
 #[clap(author, version, about)]
 struct Args {
-    #[clap(subcommand)]
-    command: Option<Commands>,
-
     #[clap(short = 'c', long = "config")]
     config_path: Option<String>,
 
     #[clap(flatten)]
     verbose: clap_verbosity_flag::Verbosity,
-}
-
-#[derive(Subcommand, Debug)]
-enum Commands {
-    Check,
 }
 
 lazy_static::lazy_static! {
@@ -51,6 +45,26 @@ lazy_static::lazy_static! {
     static ref ALERTS_CONFIG: RwLock<Vec<AlertsConfig>> = RwLock::new(Vec::new());
 }
 
+fn init_pool() -> Pool {
+    // Init the connection to the postgresql
+    let manager = ConnectionManager::<PgConnection>::new(&CONFIG.database_url);
+    // This step might spam for error CONFIG.database_max_connection of times, this is normal.
+    match r2d2::Pool::builder()
+        .max_size(CONFIG.database_max_connection)
+        .min_idle(Some((10 * CONFIG.database_max_connection) / 100))
+        .build(manager)
+    {
+        Ok(pool) => {
+            info!("R2D2 PostgreSQL pool created");
+            pool
+        }
+        Err(e) => {
+            error!("Failed to create db pool: {}", e);
+            std::process::exit(1);
+        }
+    }
+}
+
 #[tokio::main]
 async fn main() -> std::io::Result<()> {
     let args = Args::parse();
@@ -64,30 +78,38 @@ async fn main() -> std::io::Result<()> {
         .filter_module("sproot", args.verbose.log_level_filter())
         .init();
 
-    // Init the connection to the postgresql
-    let manager = ConnectionManager::<PgConnection>::new(&CONFIG.database_url);
-    // This step might spam for error CONFIG.database_max_connection of times, this is normal.
-    let pool = match r2d2::Pool::builder()
-        .max_size(CONFIG.database_max_connection)
-        .min_idle(Some((10 * CONFIG.database_max_connection) / 100))
-        .build(manager)
-    {
-        Ok(pool) => {
-            info!("R2D2 PostgreSQL pool created");
-            pool
-        }
-        Err(e) => {
-            error!("Failed to create db pool: {}", e);
-            std::process::exit(1);
-        }
+    let pool = init_pool();
+
+    let ws_handler = WsHandler {
+        query: "*",
+        table: "alerts",
+        msg_error: msg_err_handler,
+        msg_ok: msg_ok_database,
+        pool: &pool,
     };
 
-    // Dispatch subcommands
-    if let Some(Commands::Check) = &args.command {
-        check::dry_run(pool);
-        std::process::exit(0);
-    }
+    // Check if the SMTP server host is "ok"
+    test_smtp_transport();
 
-    let sp_alerts = SpAlerts::default(&pool);
-    sp_alerts.prepare().then(|_| sp_alerts.serve()).await
+    let monitor = Monitor::default(&pool);
+    // Run the foreach loop over each alarms and start monitoring them.
+    monitor.oneshot();
+
+    let mut count = 0u8;
+    loop {
+        // Create and start listening on the Websocket
+        if let Err(err) = ws_handler.listen().await {
+            error!("AlertSource::Database: stream error: {}", err);
+        }
+
+        error!("Main loop: attempt to recover the CDC connection, waiting 5s");
+        // Avoid spamming CPU in case of crash loop
+        thread::sleep(Duration::from_secs(5));
+
+        count += 1;
+        if count >= 3 {
+            error!("Main loop: boot loop, stopping...");
+            std::process::exit(1);
+        }
+    }
 }
